Added watering frequency adjustment report.
This shows which plants have not needed water in longer than the
water_freq interval several times in the last month, so I can go in and
tweak the water_freq to be less frequent for them.

Next step for this is to also include plants that have needed watering
_more_ often than water_freq recenly, so I can go in and adjust that to
make them show up on the watering list more often.